Convert package object to top level functions in Scala 3
